agents-playground/oreilly-agent-mvp/pyproject.toml
[project]
name = "agent-mvp"
version = "0.1.0"
description = "O'Reilly AI Agents MVP: Issue Triage + Implementation Draft Pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "O'Reilly Training" }]
keywords = ["ai", "agents", "langgraph", "crewai", "mcp"]

dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "crewai>=0.80.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
]

[project.scripts]
agent-mvp = "agent_mvp.pipeline.run_once:main"
agent-watcher = "agent_mvp.watcher.folder_watcher:main"
agent-menu = "agent_mvp.cli.interactive_menu:main"
agent-mcp = "agent_mvp.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_mvp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
